+++ package.json
@@ -1,57 +1,57 @@
 {
-  "name": "api",
-  "version": "1.0.0",
-  "description": "",
-  "main": "index.js",
-  "scripts": {
-    "start": "nodemon -L",
-    "resetProducts": "node ./src/controllers/DB/resetProductsController.js",
-    "resetUsers": "node ./src/controllers/DB/resetUsersController.js",
-    "test": "mocha --recursive --timeout 10000 --exit",
-    "testDB": "mocha -g 'DATABASE CONNECTIONS' --recursive --timeout 10000 --exit",
-    "testEnv": "mocha -g 'ENVIRONMENT VARIABLES' --recursive --timeout 10000 --exit",
-    "testControllers": "mocha -g 'CONTROLLERS' --recursive --timeout 10000 --exit",
-    "testCartsControllers": "mocha -g 'Carts Controllers' --recursive --timeout 10000 --exit",
-    "testUsersControllers": "mocha -g 'Users Controllers' --recursive --timeout 10000 --exit"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
-  "dependencies": {
-    "axios": "^1.4.0",
-    "bcrypt": "^5.1.0",
-    "body-parser": "^1.20.2",
-    "cloudinary": "^1.38.0",
-    "cookie-parser": "^1.4.6",
-    "cors": "^2.8.5",
-    "dotenv": "^16.3.1",
-    "express": "^4.18.2",
-    "express-openid-connect": "^2.16.0",
-    "express-rate-limit": "^6.7.1",
-    "jsonwebtoken": "^9.0.1",
-    "mercadopago": "^1.5.17",
-    "moment": "^2.29.4",
-    "morgan": "^1.10.0",
-    "multer": "^1.4.5-lts.1",
-    "nodemailer": "^6.9.4",
-    "path": "^0.12.7",
-    "pg": "^8.11.1",
-    "pg-hstore": "^2.3.4",
-    "sanitize-filename": "^1.6.3",
-    "sequelize": "^6.32.1",
-    "validator": "^13.9.0"
-  },
-  "devDependencies": {
-    "chai": "^4.2.0",
-    "eslint": "^8.46.0",
-    "mocha": "^10.2.0",
-    "nodemon": "^2.0.6",
-    "standard": "^17.1.0",
-    "supertest": "^6.0.1",
-    "supertest-session": "^4.1.0"
-  },
-  "http.proxySupport": "off",
-  "eslintConfig": {
-    "extends": "standard"
-  }
+    "name": "api",
+    "version": "1.0.0",
+    "description": "",
+    "main": "index.js",
+    "scripts": {
+        "start": "nodemon -L",
+        "resetProducts": "node ./src/controllers/DB/resetProductsController.js",
+        "resetUsers": "node ./src/controllers/DB/resetUsersController.js",
+        "test": "mocha --recursive --timeout 10000 --exit",
+        "testDB": "mocha -g 'DATABASE CONNECTIONS' --recursive --timeout 10000 --exit",
+        "testEnv": "mocha -g 'ENVIRONMENT VARIABLES' --recursive --timeout 10000 --exit",
+        "testControllers": "mocha -g 'CONTROLLERS' --recursive --timeout 10000 --exit",
+        "testCartsControllers": "mocha -g 'Carts Controllers' --recursive --timeout 10000 --exit",
+        "testUsersControllers": "mocha -g 'Users Controllers' --recursive --timeout 10000 --exit"
+    },
+    "keywords": [],
+    "author": "",
+    "license": "ISC",
+    "dependencies": {
+        "axios": "^1.4.0",
+        "bcrypt": "^5.1.0",
+        "body-parser": "^1.20.2",
+        "cloudinary": "^1.38.0",
+        "cookie-parser": "^1.4.6",
+        "cors": "^2.8.5",
+        "dotenv": "^16.3.1",
+        "express": "^4.18.2",
+        "express-openid-connect": "^2.16.0",
+        "express-rate-limit": "^6.7.1",
+        "jsonwebtoken": "^9.0.1",
+        "mercadopago": "^1.5.17",
+        "moment": "^2.29.4",
+        "morgan": "^1.10.0",
+        "multer": "^1.4.5-lts.1",
+        "nodemailer": "^6.9.4",
+        "path": "^0.12.7",
+        "pg": "^8.11.1",
+        "pg-hstore": "^2.3.4",
+        "sanitize-filename": "^1.6.3",
+        "sequelize": "^6.32.1",
+        "validator": "^13.9.0"
+    },
+    "devDependencies": {
+        "chai": "^4.2.0",
+        "eslint": "^8.46.0",
+        "mocha": "^10.2.0",
+        "nodemon": "^2.0.6",
+        "standard": "^17.1.0",
+        "supertest": "^6.0.1",
+        "supertest-session": "^4.1.0"
+    },
+    "http.proxySupport": "off",
+    "eslintConfig": {
+        "extends": "standard"
+    }
 }
