+++ src/controllers/Users/putUserController.js
@@ -1,4 +1,6 @@
 const { Users } = require("../../db");
+
+
 const putUserController = async (id, user) => {
     const userToEdit = await Users.findByPk(id);
 
@@ -6,11 +8,14 @@ const putUserController = async (id, user) => {
 
     console.log(userToEdit.dataValues.authZero);
 
-    if (user.email || user.password && userToEdit.dataValues.authZero) throw Error("No es posible actualizar un usuario que se registró por medio de authZero");
+    if (user.email || user.password && userToEdit.dataValues.authZero)
+        throw Error("No es posible actualizar un usuario que se registró" +
+            "por medio de authZero");
 
     await userToEdit.update(user);
     delete userToEdit.dataValues.password;
     return userToEdit.dataValues;
 };
+
 
 module.exports = putUserController;
