+++ src/services/index.js
@@ -2,11 +2,13 @@ const uploadImage = require('./cloudinary');
 const createToken = require('./jwt');
 const transporter = require('./nodemailer');
 const mercadopago = require('./mercadopago');
+const authZero    = require('./authZero');
 
 
 module.exports = {
     uploadImage,
     createToken,
     transporter,
-    mercadopago
+    mercadopago,
+    authZero
 };
